Rename MainLayout, build header
