+++ frontend/src/router/index.jsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "./pages/App";
-import MainLayout from "./layouts/Main";
+import MainLayout from "./layouts/MainLayout";
 import ErrorPage from "./pages/Error";
 import Foo from "./pages/Foo";
 import Bar from "./pages/Bar";
