+++ frontend/src/router/layouts/MainLayout.scss
@@ -1,0 +1,9 @@
+.layout.main {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+
+  & > .container {
+    flex: 1;
+  }
+}
